Add BaseDefinitionLoader and use it ;)
