+++ src/NullDevelopment/Skeleton/SourceCode/DefinitionLoader/SingleValueObjectLoader.php
@@ -6,13 +6,6 @@ namespace NullDevelopment\Skeleton\SourceCode\DefinitionLoader;
 
 use NullDevelopment\PhpStructure\DataTypeName\ClassName;
 use NullDevelopment\Skeleton\SourceCode\Definition\SingleValueObject;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\ConstantCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\ConstructorMethodFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\InterfaceNameCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\MethodCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\PropertyCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\TraitNameCollectionFactory;
 use NullDevelopment\Skeleton\SourceCode\Method\DeserializeMethod;
 use NullDevelopment\Skeleton\SourceCode\Method\GetterMethod;
 use NullDevelopment\Skeleton\SourceCode\Method\HasPropertyMethod;
@@ -24,42 +17,8 @@ use NullDevelopment\Skeleton\SourceCode\Method\ToStringMethod;
  * @see SingleValueObjectLoaderTest
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
-class SingleValueObjectLoader implements DefinitionLoader
+class SingleValueObjectLoader extends BaseDefinitionLoader
 {
-    /** @var InterfaceNameCollectionFactory */
-    private $interfaceNameCollectionFactory;
-
-    /** @var TraitNameCollectionFactory */
-    private $traitNameCollectionFactory;
-
-    /** @var ConstantCollectionFactory */
-    private $constantCollectionFactory;
-
-    /** @var ConstructorMethodFactory */
-    private $constructorMethodFactory;
-
-    /** @var PropertyCollectionFactory */
-    private $propertyCollectionFactory;
-
-    /** @var MethodCollectionFactory */
-    private $methodCollectionFactory;
-
-    public function __construct(
-        InterfaceNameCollectionFactory $interfaceNameCollectionFactory,
-        TraitNameCollectionFactory $traitNameCollectionFactory,
-        ConstantCollectionFactory $constantCollectionFactory,
-        ConstructorMethodFactory $constructorMethodFactory,
-        PropertyCollectionFactory $propertyCollectionFactory,
-        MethodCollectionFactory $methodCollectionFactory
-    ) {
-        $this->interfaceNameCollectionFactory = $interfaceNameCollectionFactory;
-        $this->traitNameCollectionFactory     = $traitNameCollectionFactory;
-        $this->constantCollectionFactory      = $constantCollectionFactory;
-        $this->constructorMethodFactory       = $constructorMethodFactory;
-        $this->propertyCollectionFactory      = $propertyCollectionFactory;
-        $this->methodCollectionFactory        = $methodCollectionFactory;
-    }
-
     public function supports(array $input): bool
     {
         if ('SingleValueObject' === $input['type']) {
@@ -123,17 +82,5 @@ class SingleValueObjectLoader implements DefinitionLoader
             'methods'     => [],
             'constructor' => [],
         ];
-    }
-
-    private function extractParent(array $data): ?ClassName
-    {
-        if (null === $data['parent']) {
-            return null;
-        }
-        if (true === is_array($data['parent'])) {
-            return ClassName::create($data['parent']['instanceOf'], $data['parent']['alias']);
-        }
-
-        return ClassName::create($data['parent']);
     }
 }
