+++ src/NullDevelopment/SkeletonBroadwayExtension/EventSourcedAggregateRoot/SourceCode/EventSourcedAggregateRootLoader.php
@@ -5,54 +5,14 @@ declare(strict_types=1);
 namespace NullDevelopment\SkeletonBroadwayExtension\EventSourcedAggregateRoot\SourceCode;
 
 use NullDevelopment\PhpStructure\DataTypeName\ClassName;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\ConstantCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\ConstructorMethodFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\InterfaceNameCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\MethodCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\PropertyCollectionFactory;
-use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\Factory\TraitNameCollectionFactory;
+use NullDevelopment\Skeleton\SourceCode\DefinitionLoader\BaseDefinitionLoader;
 
 /**
  * @see EventSourcedAggregateRootLoaderSpec
  * @see EventSourcedAggregateRootLoaderTest
  */
-class EventSourcedAggregateRootLoader implements DefinitionLoader
+class EventSourcedAggregateRootLoader extends BaseDefinitionLoader
 {
-    /** @var InterfaceNameCollectionFactory */
-    private $interfaceNameCollectionFactory;
-
-    /** @var TraitNameCollectionFactory */
-    private $traitNameCollectionFactory;
-
-    /** @var ConstantCollectionFactory */
-    private $constantCollectionFactory;
-
-    /** @var ConstructorMethodFactory */
-    private $constructorMethodFactory;
-
-    /** @var PropertyCollectionFactory */
-    private $propertyCollectionFactory;
-
-    /** @var MethodCollectionFactory */
-    private $methodCollectionFactory;
-
-    public function __construct(
-        InterfaceNameCollectionFactory $interfaceNameCollectionFactory,
-        TraitNameCollectionFactory $traitNameCollectionFactory,
-        ConstantCollectionFactory $constantCollectionFactory,
-        ConstructorMethodFactory $constructorMethodFactory,
-        PropertyCollectionFactory $propertyCollectionFactory,
-        MethodCollectionFactory $methodCollectionFactory
-    ) {
-        $this->interfaceNameCollectionFactory = $interfaceNameCollectionFactory;
-        $this->traitNameCollectionFactory     = $traitNameCollectionFactory;
-        $this->constantCollectionFactory      = $constantCollectionFactory;
-        $this->constructorMethodFactory       = $constructorMethodFactory;
-        $this->propertyCollectionFactory      = $propertyCollectionFactory;
-        $this->methodCollectionFactory        = $methodCollectionFactory;
-    }
-
     public function supports(array $input): bool
     {
         if ('BroadwayEventSourcedAggregateRoot' === $input['type']) {
@@ -103,17 +63,5 @@ class EventSourcedAggregateRootLoader implements DefinitionLoader
             'methods'     => [],
             'constructor' => [],
         ];
-    }
-
-    private function extractParent(array $data): ?ClassName
-    {
-        if (null === $data['parent']) {
-            return null;
-        }
-        if (true === is_array($data['parent'])) {
-            return ClassName::create($data['parent']['instanceOf'], $data['parent']['alias']);
-        }
-
-        return ClassName::create($data['parent']);
     }
 }
